Round factor and new dialog at run

+++ src/QLearning/State.java
@@ -13,8 +13,8 @@ These variables are then rounded in order to create an approximate state.
 
 public class State {
     // State features
-    private double worldAngle; // Torso angle relative to World
-    private ArrayList<Double> jointAngles = new ArrayList<>(); // Angles of joints
+    private int worldAngle; // Torso angle relative to World
+    private ArrayList<Integer> jointAngles = new ArrayList<>(); // Angles of joints
     private static ArrayList<JointAction> actions = new ArrayList<>(); // List containing possible actions. Static since actions are the same for all states
     public static int roundFactor = 15; // Round factor - the higher the number the lower precision in states // TODO WHAT VALUE!!
 
@@ -22,9 +22,9 @@ public class State {
         // Constructor for a new state with the walker as input.
         // The jointangles from the walker are rounded and then added to the state in order to make the state approximate of the "actual" state
         for (RevoluteJoint j : walker.joints) {
-            jointAngles.add((double) (Math.round(Math.toDegrees(j.getJointAngle()) / roundFactor)));
+            jointAngles.add((int) (Math.round(Math.toDegrees(j.getJointAngle()) / roundFactor)));
         }
-        this.worldAngle = (double) (Math.round(Math.toDegrees(walker.getRelativeAngle()) / roundFactor));
+        this.worldAngle = (int) (Math.round(Math.toDegrees(walker.getRelativeAngle()) / roundFactor));
     }
 
     public static void fillActions() {
@@ -82,6 +82,7 @@ public class State {
     }
 
     public static String getTheoreticalNumberOfStates() {
+
 
         int hipInterval = Math.round(55) / roundFactor;
         int kneeInterval = Math.round(150) / roundFactor;
